refactor: inline hook processing, remove runHooks and resumeGenerator
executeRerender and resumeInstance now process hook descriptors inline
via processOneDescriptor, replacing the separate runHooks and
resumeGenerator functions. This eliminates two layers of indirection —
the component generator's hook yields are processed directly in the
lifecycle functions.
